[MOD/#1114] 알림 아이템 함수화
