+++ feature/notification/src/main/java/org/sopt/official/feature/notification/all/NotificationActivity.kt
@@ -26,29 +26,22 @@ package org.sopt.official.feature.notification.all
 
 import android.content.Context
 import android.content.Intent
-import android.icu.text.DateFormat
-import android.icu.text.SimpleDateFormat
-import android.icu.util.TimeZone
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
@@ -61,21 +54,18 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.paging.compose.collectAsLazyPagingItems
 import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.EntryPointAccessors
-import java.util.Date
-import java.util.Locale
 import org.sopt.official.common.navigator.NavigatorEntryPoint
 import org.sopt.official.common.view.toast
 import org.sopt.official.designsystem.Orange400
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.feature.notification.R
 import org.sopt.official.feature.notification.all.component.NotificationCategoryChip
+import org.sopt.official.feature.notification.all.component.NotificationInfoItem
 
 @AndroidEntryPoint
 class NotificationActivity : AppCompatActivity() {
@@ -162,62 +152,18 @@ class NotificationActivity : AppCompatActivity() {
                         if (notifications.itemCount > 0) {
                             LazyColumn {
                                 items(notifications.itemCount) {
-                                    val item = notifications[it]
-                                    Column(
-                                        modifier = Modifier
-                                            .fillMaxWidth()
-                                            .height(100.dp)
-                                            .clickable {
-                                                if (item?.notificationId == null) {
-                                                    toast("문제가 발생했습니다.")
-                                                } else {
-                                                    context.startActivity(navigator.getNotificationDetailActivityIntent(item.notificationId))
-                                                }
+                                    val notification = notifications[it]
+
+                                    NotificationInfoItem(
+                                        notification = notification,
+                                        onCLick = {
+                                            if (notification?.notificationId == null) {
+                                                toast("문제가 발생했습니다.")
+                                            } else {
+                                                context.startActivity(navigator.getNotificationDetailActivityIntent(notification.notificationId))
                                             }
-                                            .background(
-                                                if (item?.isRead == true) {
-                                                    SoptTheme.colors.onSurface800
-                                                } else {
-                                                    SoptTheme.colors.background
-                                                }
-                                            )
-                                            .padding(
-                                                horizontal = 20.dp,
-                                                vertical = 16.dp
-                                            )
-                                    ) {
-                                        Row(
-                                            modifier = Modifier.fillMaxWidth(),
-                                            horizontalArrangement = Arrangement.SpaceBetween,
-                                            verticalAlignment = Alignment.CenterVertically
-                                        ) {
-                                            Text(
-                                                text = item?.title.orEmpty(),
-                                                style = SoptTheme.typography.body16M,
-                                                color = SoptTheme.colors.onSurface30,
-                                                modifier = Modifier
-                                                    .weight(1f)
-                                                    .widthIn(max = 250.dp),
-                                                maxLines = 1,
-                                                overflow = TextOverflow.Ellipsis
-                                            )
-                                            Text(
-                                                text = item?.createdAt.orEmpty().convertToTimesAgo(),
-                                                style = SoptTheme.typography.body13M.copy(fontSize = 12.sp),
-                                                color = SoptTheme.colors.onSurface100
-                                            )
                                         }
-                                        Spacer(modifier = Modifier.height(8.dp))
-                                        Text(
-                                            text = item?.content.orEmpty(),
-                                            style = SoptTheme.typography.body16M,
-                                            color = SoptTheme.colors.onSurface400,
-                                            maxLines = 2,
-                                            overflow = TextOverflow.Ellipsis,
-                                            modifier = Modifier.weight(1f)
-                                        )
-                                    }
-                                    HorizontalDivider(color = SoptTheme.colors.onSurface600)
+                                    )
                                 }
                             }
                         } else {
@@ -245,39 +191,10 @@ class NotificationActivity : AppCompatActivity() {
         }
     }
 
-    private fun String.convertToTimesAgo(): String {
-        if (this.isEmpty()) return ""
-
-        val dateFormat: DateFormat = SimpleDateFormat(
-            "yyyy-MM-dd'T'HH:mm:ss.SSSSSS",
-            Locale.KOREA
-        )
-        dateFormat.timeZone = TimeZone.getTimeZone("Asia/Seoul")
-
-        val currentDate = Date()
-        val receivedDate = dateFormat.parse(this)
-        val diffInMillis = currentDate.time - receivedDate.time
-
-        val diffInDays = diffInMillis / ONE_DAY_IN_MILLISECONDS
-        val diffInHours = diffInMillis / ONE_HOUR_IN_MILLISECONDS
-        val diffInMinutes = diffInMillis / ONE_MINUTE_IN_MILLISECONDS
-
-        return when {
-            diffInDays >= 1 -> "${diffInDays}일 전"
-            diffInHours >= 1 -> "${diffInHours}시간 전"
-            diffInMinutes >= 1 -> "${diffInMinutes}분 전"
-            else -> "방금"
-        }
-    }
-
     companion object {
         fun newInstance(context: Context) = Intent(
             context,
             NotificationActivity::class.java
         )
-
-        const val ONE_DAY_IN_MILLISECONDS = 86400000L
-        const val ONE_HOUR_IN_MILLISECONDS = 3600000L
-        const val ONE_MINUTE_IN_MILLISECONDS = 60000L
     }
 }
